Re-split code into trait, recover PrimaryRead connection
Fix Psalm issues
