+++ src/PrimaryReadReplicaConnection.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Facile\DoctrineMySQLComeBack\Doctrine\DBAL;
+
+class PrimaryReadReplicaConnection extends \Doctrine\DBAL\Connections\PrimaryReadReplicaConnection
+{
+    use ConnectionTrait;
+}
